feat: fix strategy execution engine, standardize configs, and fix paper trading simulation
- Standardized `generate_signals` output across all strategies to conform to a universal contract (added `entry_config`).
- Fixed live stop-loss and take-profit execution in `instances/runner.py` to evaluate against static entry values stored in `active_trade` instead of dynamic/receding values from `exit_config` on each tick.
- Implemented full-fidelity mathematical simulation of paper trading balance and trade PnL to resolve 0.0 values in dry-runs.
- Routed paper trading exclusively through `app/paper_routes.py` (fixed NameError, resolved duplicate route, and enforced per-user API key isolation).
- Addressed BUG-F2 through BUG-F7 (added `@dataclass` on BacktestTrade, added None guards in backtest runner, fixed UI profile 'balanced' mismatch, updated entry cost calculations, and schema type mismatch).
- Fixed live-update polling for paper trading charts in `testing_paper.html`.
- Updated test suite endpoints to align with correct kill-switch routes.

+++ api/backtests.py
@@ -34,7 +34,7 @@ class RunBacktestRequest(BaseModel):
     timeframe: str | None = None
     mode: str | None = None
     profile: str | None = None
-    activation: float | None = None
+    activation: int | None = None
     offset: int | None = None
     leverage: int | None = None
     tick_mode: int = Field(default=1, ge=1, le=28)  # 1=OHLC, 4=basic, 28=high
